Perf - use Lazy<bool> instead of encuring the cost of IsEntitySet every time
It won't change once the EntityDataModel is built, however we don't know until all the entity sets are registered whether a property type is navigable or not.

e.g. if TypeA has a property of type TypeB and TypeB is added as an entity set after TypeA, when the properties are verified for TypeA, the TypeB property is not a known entity set and therefore not navigable.

Although more complex in the implementation it avoids the prospect of having to try and register entity sets in order to make sure that referenced properties which are entity sets are registered first and also that still won't necessary work if there are circular referenes between types all of which are entity sets.