Added a generic version of ConstantNode
This makes it possible to avoid boxing/unboxing the value if it's a value type.

Note - The QueryNode will still always reference the base class ConstantNode so it requires casting the ConstantNode to ConstantNode<T> based upon the ConstantNode.EdmType otherwise the value will be boxed on demand by the getter in the base ConstantNode class.